backend/radio/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
ffmpeg-python==0.2.0
future==0.18.3
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.3.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
